requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.0
django-embed-video==1.4.0
djangorestframework==3.12.4
djongo==1.3.4
idna==2.10
Pillow==8.2.0
pymongo==3.11.3
pytz==2021.1
requests==2.25.1
sqlparse==0.2.4
urllib3==1.26.4
